+++ apps/handover/CHANGELOG.md
@@ -1,3 +1,6 @@
+## Version 0.0.11
+Inject process.env to bundle
+
 ## Version 0.0.10
 Bump fusion-framework packages