+++ apps/mechanicalcompletion/CHANGELOG.md
@@ -1,3 +1,6 @@
+## Version 0.0.5
+Inject process.env to bundle
+
 ## Version 0.0.4
 Bump fusion-framework packages